Audio management interface (list / delete / play), youtube downloader + progress, textoverlay perfect adjustment for ifttt, ifttt multi condition statement parsing, diskfree sensor

+++ client/client/actors/ledstrip.js
@@ -28,7 +28,7 @@ exports.exposed = function()
                 name: "title",
                 isOptional: false,
                 dataType: "string",
-                notes: "name of the song or full link to youtube video that should be played as lightshow"
+                notes: "name of the song that should be played as lightshow"
             }]
         },
         singleColor: {
@@ -57,11 +57,19 @@ exports.allOff = function()
 {
     logger.info("disabling ledstrip");
 
-    if (spawned)
+    if (spawned && spawned.pid)
     {
         logger.info("killing pid group", spawned.pid);
         //kill subprocesses aswell
-        process.kill(-spawned.pid);
+        try
+        {
+            process.kill(-spawned.pid);
+        }
+        catch (err)
+        {
+            logger.error("could not kill process", err);
+        }
+
         spawned = null;
     }
 
@@ -102,24 +110,6 @@ exports.randomColor = function()
 
 exports.lightshow = function(title)
 {
-    if (title.indexOf("http") === 0)
-    {
-        return youtube.download(title, function onout(text)
-        {
-            logger.info(text);
-        },
-        function onclose(code, fileName)
-        {
-            logger.info("Done with response code: " + code + " and file " + fileName);
-
-            if (code === 0 && fileName)
-            {
-                logger.info("Playing " + fileName);
-                exports.lightshow(fileName);
-            }
-        });
-    }
-
     title = title || "song.mp3";
     title = title.replace("..", "");
     title = config.mediaBasePath + "/" + title;
